workers: price-reporter: reorganize price reporter modules for external executor
